+++ R/runDebug.R
@@ -1,7 +1,7 @@
 ######################################################################################
 #' @title runDebug 
 #'
-#' @description runDebug is an internal function used by GRD and superbPlot
+#' @description runDebug is an internal function used by GRD and superb
 #'      to help in debugging the functions. It assigns in the global environment
 #'      the variables that are local to a function so that they become visible.
 #'